Submitter class hierarchy
